--- a/apps/vscode/webpack.prod.config.js
+++ b/apps/vscode/webpack.prod.config.js
@@ -1,0 +1,5 @@
+module.exports = function transformWebpack(config) {
+  // the default node webpack config has minimize: false when optimization is set to true
+  delete config.optimization;
+  return config;
+};