--- a/apps/vscode/webpack.prod.config.js
+++ b/apps/vscode/webpack.prod.config.js
@@ -1,5 +1,14 @@
 module.exports = function transformWebpack(config) {
   // the default node webpack config has minimize: false when optimization is set to true
-  delete config.optimization;
+  config.output.filename = '[name].js';
+  config.optimization = {
+    minimize: true,
+    concatenateModules: true,
+    sideEffects: false,
+    splitChunks: {
+      chunks: 'all',
+    },
+  };
+
   return config;
 };