--- a/libs/server/src/lib/abstract-tree-provider.ts
+++ b/libs/server/src/lib/abstract-tree-provider.ts
@@ -7,12 +7,12 @@ import {
 } from 'vscode';
 
 export abstract class AbstractTreeProvider<Item extends TreeItem>
-  implements TreeDataProvider<Item> {
-  private readonly _onDidChangeTreeData: EventEmitter<
-    Item | undefined
-  > = new EventEmitter();
-  readonly onDidChangeTreeData: Event<Item | undefined> = this
-    ._onDidChangeTreeData.event;
+  implements TreeDataProvider<Item>
+{
+  private readonly _onDidChangeTreeData: EventEmitter<Item | undefined> =
+    new EventEmitter();
+  readonly onDidChangeTreeData: Event<Item | undefined> =
+    this._onDidChangeTreeData.event;
 
   refresh(): void {
     this._onDidChangeTreeData.fire(undefined);
@@ -22,7 +22,6 @@ export abstract class AbstractTreeProvider<Item extends TreeItem>
     return element;
   }
 
-  abstract getParent(element: Item): Item | null | undefined;
-
+  abstract getParent(element: Item): ProviderResult<Item | null | undefined>;
   abstract getChildren(element?: Item): ProviderResult<Item[]>;
 }