--- a/libs/vscode/nx-workspace/src/lib/get-nx-config.ts
+++ b/libs/vscode/nx-workspace/src/lib/get-nx-config.ts
@@ -3,12 +3,14 @@ import { NxJsonConfiguration } from '@nrwl/devkit';
 import { join } from 'path';
 import { getNxWorkspacePackageFileUtils } from './get-nx-workspace-package';
 
-export function getNxConfig(baseDir: string): NxJsonConfiguration {
+export async function getNxConfig(
+  baseDir: string
+): Promise<NxJsonConfiguration> {
   try {
     let cachedNxJson = cacheJson('nx.json', baseDir).json;
 
     if (!cachedNxJson) {
-      const nxJson = getNxWorkspacePackageFileUtils().readNxJson(
+      const nxJson = (await getNxWorkspacePackageFileUtils()).readNxJson(
         join(baseDir, 'nx.json')
       );