--- a/libs/vscode/nx-workspace/src/lib/get-nx-workspace-package.ts
+++ b/libs/vscode/nx-workspace/src/lib/get-nx-workspace-package.ts
@@ -1,24 +1,35 @@
 import { WorkspaceConfigurationStore } from '@nx-console/vscode/configuration';
 import { dirname, join } from 'path';
-import type * as NxWorkspacePackage from '@nrwl/workspace';
+import * as NxWorkspaceFileUtils from '@nrwl/workspace/src/core/file-utils';
+import { getOutputChannel } from '@nx-console/server';
+
 /**
  * Get the local installed version of @nrwl/workspace
  */
-export function getNxWorkspacePackageFileUtils(): typeof NxWorkspacePackage {
+export async function getNxWorkspacePackageFileUtils(): Promise<
+  typeof NxWorkspaceFileUtils
+> {
   const workspacePath = dirname(
     WorkspaceConfigurationStore.instance.get('nxWorkspaceJsonPath', '')
   );
 
-  // webpack hacks..
-  return eval('require')(
-    join(
-      workspacePath,
-      'node_modules',
-      '@nrwl',
-      'workspace',
-      'src',
-      'core',
-      'file-utils'
-    )
-  );
+  try {
+    return import(
+      /*webpackIgnore: true*/
+      join(
+        workspacePath,
+        'node_modules',
+        '@nrwl',
+        'workspace',
+        'src',
+        'core',
+        'file-utils'
+      )
+    );
+  } catch (err) {
+    getOutputChannel().appendLine(
+      `Error loading @nrwl/workspace from workspace. Falling back to extension dependency`
+    );
+    return NxWorkspaceFileUtils;
+  }
 }