--- a/package.json
+++ b/package.json
@@ -2,6 +2,7 @@
   "name": "nx-console",
   "version": "17.10.0",
   "publisher": "nrwl",
+  "sideEffects": false,
   "author": {
     "name": "Narwhal Technologies Inc",
     "email": "hello@nrwl.io"