## macOS
17.10.0 (average time (load code + call activate) - `769ms`)
```
| Extension                             | Eager | Load Code | Call Activate | Finish Activate | Event                                                 | By                                    |
| ------------------------------------- | ----- | --------- | ------------- | --------------- | ----------------------------------------------------- | ------------------------------------- |
| nrwl.angular-console                  | false | 309       | 499           | 0               | onStartupFinished                                     | nrwl.angular-console                  |
```
```
| Extension                             | Eager | Load Code | Call Activate | Finish Activate | Event                                                 | By                                    |
| ------------------------------------- | ----- | --------- | ------------- | --------------- | ----------------------------------------------------- | ------------------------------------- |
| nrwl.angular-console                  | false | 306       | 466           | 1               | onStartupFinished                                     | nrwl.angular-console                  |
```
```
| Extension                             | Eager | Load Code | Call Activate | Finish Activate | Event                                                 | By                                    |
| ------------------------------------- | ----- | --------- | ------------- | --------------- | ----------------------------------------------------- | ------------------------------------- |
onStartupFinished                                     | ms-vscode-remote.remote-wsl           |
| nrwl.angular-console                  | false | 291       | 437           | 1               | 
```

With these changes (`446ms`)
```
| Extension                             | Eager | Load Code | Call Activate | Finish Activate | Event                                                 | By                                    |
| ------------------------------------- | ----- | --------- | ------------- | --------------- | ----------------------------------------------------- | ------------------------------------- |
| nrwl.angular-console                  | false | 410       | 28            | 22              | onStartupFinished                                     | nrwl.angular-console                  |
```
```
| Extension                             | Eager | Load Code | Call Activate | Finish Activate | Event                                                 | By                                    |
| ------------------------------------- | ----- | --------- | ------------- | --------------- | ----------------------------------------------------- | ------------------------------------- |
| nrwl.angular-console                  | false | 428       | 28            | 18              | onStartupFinished
```
```
| Extension                             | Eager | Load Code | Call Activate | Finish Activate | Event                                                 | By                                    |
| ------------------------------------- | ----- | --------- | ------------- | --------------- | ----------------------------------------------------- | ------------------------------------- |
| nrwl.angular-console                  | false | 415       | 29            | 20              | onStartupFinished                                     | nrwl.angular-console                  |

```

Stats
<img width="698" alt="image" src="https://user-images.githubusercontent.com/4332460/132909092-5f3db2e1-4ff7-4b9f-8c83-5718ee989dab.png">


## Windows
17.10.0 (average time (load code + call activate) - `1039ms`)
```
| Extension                               | Eager | Load Code | Call Activate | Finish Activate | Event                                                 | By                                      |
| --------------------------------------- | ----- | --------- | ------------- | --------------- | ----------------------------------------------------- | --------------------------------------- |
| nrwl.angular-console                    | false | 301       | 801           | 3               | onStartupFinished                                     | nrwl.angular-console                    |

```
```
| Extension                               | Eager | Load Code | Call Activate | Finish Activate | Event                                                 | By                                      |
| --------------------------------------- | ----- | --------- | ------------- | --------------- | ----------------------------------------------------- | --------------------------------------- |
| nrwl.angular-console                    | false | 164       | 920           | 0               | onStartupFinished                                     | nrwl.angular-console                    |

```
```
| Extension                               | Eager | Load Code | Call Activate | Finish Activate | Event                                                 | By                                      |
| --------------------------------------- | ----- | --------- | ------------- | --------------- | ----------------------------------------------------- | --------------------------------------- |
| nrwl.angular-console                    | false | 209       | 723           | 15              | onStartupFinished                                     | nrwl.angular-console                    |
```

With these changes (on windows average time - `271ms`)
```
| Extension                               | Eager | Load Code | Call Activate | Finish Activate | Event                                                 | By                                      |
| --------------------------------------- | ----- | --------- | ------------- | --------------- | ----------------------------------------------------- | --------------------------------------- |
| nrwl.angular-console                    | false | 252       | 57            | 46              | onStartupFinished                                     | nrwl.angular-console                    |

```
```
| Extension                               | Eager | Load Code | Call Activate | Finish Activate | Event                                                 | By                                      |
| --------------------------------------- | ----- | --------- | ------------- | --------------- | ----------------------------------------------------- | --------------------------------------- |
| nrwl.angular-console                    | false | 270       | 2             | 31              | onStartupFinished                                     | nrwl.angular-console                    |
```
```
| Extension                               | Eager | Load Code | Call Activate | Finish Activate | Event                                                 | By                                      |
| --------------------------------------- | ----- | --------- | ------------- | --------------- | ----------------------------------------------------- | --------------------------------------- |
| nrwl.angular-console                    | false | 217       | 14            | 23              | onStartupFinished                                     | nrwl.angular-console                    |

```

stats
<img width="553" alt="image" src="https://user-images.githubusercontent.com/4332460/132909304-24371be9-3a5b-4b4f-a3bd-56d6a016ee82.png">


closes #1127
closes #1104